[tests] rename test -> tests, use native assert instead of numpy's
