+++ tests/lgdo/test_struct.py
