+++ tests/math/test_fwhm.py
@@ -1,5 +1,5 @@
+from pytest import approx
 import numpy as np
-from numpy.testing import assert_
 import pygama.math.peak_fitting as pgf
 
 
@@ -14,8 +14,8 @@ def test_mostly_gauss_fwhm():
            [0, 0, 0, 0, 0, 0, 1e-16]] # damp2
     mu, sig, hstep, htail, tau, bg0, amp = pars
     fwhm, dfwhm = pgf.radford_fwhm(sig, htail, tau, cov)
-    assert_(np.absolute(fwhm - 2.3548) < 1e-4, f'got {fwhm}')
-    assert_(np.absolute(dfwhm - 0.23548) < 1e-5, f'got {dfwhm}')
+    assert fwhm == approx(2.3548)
+    assert dfwhm == approx(0.23548)
 
 
 def test_mostly_exp_fwhm():
@@ -29,5 +29,5 @@ def test_mostly_exp_fwhm():
            [0, 0, 0, 0, 0, 0, 1e-16]] # damp2
     mu, sig, hstep, htail, tau, bg0, amp = pars
     fwhm, dfwhm = pgf.radford_fwhm(sig, htail, tau, cov)
-    assert_(np.absolute(fwhm - np.log(2)) < 1e-4, f'got {fwhm}')
-    assert_(np.absolute(dfwhm - np.log(2)/10) < 1e-5, f'got {dfwhm}')
+    assert fwhm == approx(2)
+    assert dfwhm == approx(np.log(2)/10)
